fix: fix a mistake in the code that causes a bug
There was an if statement that was supposed check if the year was
divisible by 4, 100, and 400. Instead it was checking if the year was
not divisible by 100, which I corrected.

It was also checking if the year divided by 400 strictly equalled true,
instead of checking if the year divided by 400 equalled 0. This was not
intended and I don't believe that question would have ever equalled
true, so I corrected that as well.