Nest str expressions inside of another function
Need to add more code to get this program fully functional. I can't do
that after a return statement, and the code needs to come after return
so that it use the version of str that has ignored punctuation and
capitalization. Nesting all of str's expressions inside of another
function fixes this issue.